add environ to trackers, weekly by referrer tracker, and simple middleware to display
The trackers are now taking the environ headers so they can store additional
information based on it. The new weekly referrer tracker stores hits by week,
but also groups them by url and referrer, since that's a common need to see
where the request is coming from. There is also a new piece of middleware which
just iterates through the pickles in the data_dir and prints them out to the
screen, which can be useful for debugging.


git-svn-id: http://svn.pythonpaste.org/Paste/VaingloriousEye/trunk@7820 71705e6d-80f1-0310-9d5e-e4e1c580a33d